Refactor PNLite class and DFP adapters
