+++ hybid.adapters.dfp/src/main/java/net/pubnative/lite/adapters/dfp/HyBidDFPInterstitialCustomEvent.java
@@ -1,0 +1,155 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2018 PubNative GmbH
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+package net.pubnative.lite.adapters.dfp;
+
+import android.app.Activity;
+import android.content.Context;
+import android.os.Bundle;
+import android.text.TextUtils;
+
+import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.mediation.MediationAdRequest;
+import com.google.android.gms.ads.mediation.customevent.CustomEventInterstitial;
+import com.google.android.gms.ads.mediation.customevent.CustomEventInterstitialListener;
+
+import net.pubnative.lite.sdk.HyBid;
+import net.pubnative.lite.sdk.interstitial.presenter.InterstitialPresenter;
+import net.pubnative.lite.sdk.interstitial.presenter.InterstitialPresenterFactory;
+import net.pubnative.lite.sdk.models.Ad;
+import net.pubnative.lite.sdk.utils.Logger;
+
+public class HyBidDFPInterstitialCustomEvent implements CustomEventInterstitial, InterstitialPresenter.Listener {
+    private static final String TAG = HyBidDFPInterstitialCustomEvent.class.getSimpleName();
+
+    private static final String ZONE_ID_KEY = "pn_zone_id";
+    private CustomEventInterstitialListener mInterstitialListener;
+    private InterstitialPresenter mPresenter;
+
+    @Override
+    public void requestInterstitialAd(Context context,
+                                      CustomEventInterstitialListener listener,
+                                      String serverParameter,
+                                      MediationAdRequest mediationAdRequest,
+                                      Bundle customEventExtras) {
+        if (listener == null) {
+            Logger.e(TAG, "customEventInterstitialListener is null");
+            return;
+        }
+        mInterstitialListener = listener;
+
+        if (!(context instanceof Activity)) {
+            Logger.e(TAG, "HyBid interstitial ad can only be rendered with an Activity context");
+            mInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INTERNAL_ERROR);
+            return;
+        }
+        final Activity activity = (Activity) context;
+
+        String zoneIdKey;
+        if (!TextUtils.isEmpty(HyBidDFPUtils.getZoneId(serverParameter))) {
+            zoneIdKey = HyBidDFPUtils.getZoneId(serverParameter);
+        } else if (!TextUtils.isEmpty(HyBidDFPUtils.getZoneId(customEventExtras))) {
+            zoneIdKey = HyBidDFPUtils.getZoneId(customEventExtras);
+        } else {
+            Logger.e(TAG, "Could not find zone id value in CustomEventInterstitial localExtras or serverExtras");
+            mInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INVALID_REQUEST);
+            return;
+        }
+
+        final Ad ad = HyBid.getAdCache().remove(zoneIdKey);
+        if (ad == null) {
+            Logger.e(TAG, "Could not find an ad in the cache for zone id with key " + zoneIdKey);
+            mInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_NETWORK_ERROR);
+            return;
+        }
+
+        mPresenter = new InterstitialPresenterFactory(activity).createInterstitialPresenter(ad, this);
+        if (mPresenter == null) {
+            Logger.e(TAG, "Could not create valid interstitial presenter");
+            mInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_NETWORK_ERROR);
+            return;
+        }
+
+        mPresenter.load();
+    }
+
+    @Override
+    public void onDestroy() {
+        if (mPresenter != null) {
+            mPresenter.destroy();
+            mPresenter = null;
+        }
+    }
+
+    @Override
+    public void onPause() {
+
+    }
+
+    @Override
+    public void onResume() {
+
+    }
+
+    @Override
+    public void showInterstitial() {
+        if (mPresenter != null) {
+            mPresenter.show();
+        }
+    }
+
+    @Override
+    public void onInterstitialLoaded(InterstitialPresenter interstitialPresenter) {
+        if (mInterstitialListener != null) {
+            mInterstitialListener.onAdLoaded();
+        }
+    }
+
+    @Override
+    public void onInterstitialError(InterstitialPresenter interstitialPresenter) {
+        if (mInterstitialListener != null) {
+            mInterstitialListener.onAdFailedToLoad(AdRequest.ERROR_CODE_INTERNAL_ERROR);
+        }
+    }
+
+    @Override
+    public void onInterstitialShown(InterstitialPresenter interstitialPresenter) {
+        if (mInterstitialListener != null) {
+            mInterstitialListener.onAdOpened();
+        }
+    }
+
+    @Override
+    public void onInterstitialClicked(InterstitialPresenter interstitialPresenter) {
+        if (mInterstitialListener != null) {
+            mInterstitialListener.onAdClicked();
+            mInterstitialListener.onAdLeftApplication();
+        }
+    }
+
+    @Override
+    public void onInterstitialDismissed(InterstitialPresenter interstitialPresenter) {
+        if (mInterstitialListener != null) {
+            mInterstitialListener.onAdClosed();
+        }
+    }
+}
